Move test from test_session to test_astro_session
The test uses the sherpa.astro.ui.utils Session so move it to
the correct test file.
